@mengqiy 2024-09-09
Great to see the great performance improvement!
IMO we should also confirm the throughput improvement. e.g. Running the repro in https://github.com/kubernetes/kubernetes/issues/126510#issuecomment-2314485606 should not result a huge pendingNotification backlog any more with this change.