@hakuna-matatah 2024-09-10
Also, I have noticed `pending_notifications` count with this change for 10K+ cronjobs to be around 60k, as opposed to around 500k+ based on what @mengqiy told me for 10k cronjobs.

```
# TYPE cache_pending_notifications gauge
cache_pending_notifications{name="job-controller",resources="job"} 56499
cache_pending_notifications{name="job-controller",resources="pod"} 0
cache_pending_notifications{name="ttl-after-finished-controller",resources="job"} 0
```



@alculquicondor - FYI ^^^