@alculquicondor 2024-09-11
And what was the number before the change?