@alculquicondor 2024-09-11
ok, so we effectively reduced the load to half. Great!

/lgtm
/approve