@mengqiy 2024-09-11
@alculquicondor I guess this PR doesn't have any user facing change, so I guess we may not need a release note.
@hakuna-matatah If that's the case, we can just add an empty release note block in the PR description.