@alculquicondor 2024-09-11
> This PR doesn't have any user facing change, so I guess we may not need a release note.

Performance improvements are user facing.