@alculquicondor 2024-09-11
The release note doesn't sound accurate and it should just focus on the user-facing behavior, as opposed to internal details (we have the PR description for that).

/release-note-edit

```release-note
Improve performance of the job controller when handling job update events.
```